+++ packages/snaps-execution-environments/src/offscreen/OffscreenSnapExecutor.ts
@@ -2,6 +2,7 @@ import {
   BasePostMessageStream,
   WindowPostMessageStream,
 } from '@metamask/post-message-stream';
+import { createWindow } from '@metamask/snaps-utils';
 import { JsonRpcParams, JsonRpcRequest } from '@metamask/utils';
 
 type ExecutorJob = {
@@ -69,7 +70,7 @@ export class OffscreenSnapExecutor {
    * @param frameUrl - The URL to load in the iframe.
    */
   async #initializeJob(jobId: string, frameUrl: string): Promise<ExecutorJob> {
-    const window = await this.#createWindow(frameUrl, jobId);
+    const window = await createWindow(frameUrl, jobId);
     const jobStream = new WindowPostMessageStream({
       name: 'parent',
       target: 'child',
@@ -101,66 +102,5 @@ export class OffscreenSnapExecutor {
     iframe.parentNode.removeChild(iframe);
     this.#jobs[jobId].stream.destroy();
     delete this.#jobs[jobId];
-  }
-
-  /**
-   * Creates the iframe to be used as the execution environment. This may run
-   * forever if the iframe never loads, but the promise should be wrapped in
-   * an initialization timeout in the SnapController.
-   *
-   * @param uri - The iframe URI.
-   * @param jobId - The job id.
-   * @returns A promise that resolves to the contentWindow of the iframe.
-   */
-  // TODO: Move this to a reusable utility, as it is also used in the iframe
-  // execution service.
-  async #createWindow(uri: string, jobId: string): Promise<Window> {
-    return new Promise((resolve, reject) => {
-      const iframe = document.createElement('iframe');
-      // The order of operations appears to matter for everything except this
-      // attribute. We may as well set it here.
-      iframe.setAttribute('id', jobId);
-
-      // In the past, we've had problems that appear to be symptomatic of the
-      // iframe firing the `load` event before its scripts are actually loaded,
-      // which has prevented snaps from executing properly. Therefore, we set
-      // the `src` attribute and append the iframe to the DOM before attaching
-      // the `load` listener.
-      //
-      // `load` should only fire when "all dependent resources" have been
-      // loaded, which includes scripts.
-      //
-      // MDN article for `load` event: https://developer.mozilla.org/en-US/docs/Web/API/Window/load_event
-      // Re: `load` firing twice: https://stackoverflow.com/questions/10781880/dynamically-created-iframe-triggers-onload-event-twice/15880489#15880489
-      iframe.setAttribute('src', uri);
-      document.body.appendChild(iframe);
-
-      iframe.addEventListener('load', () => {
-        if (iframe.contentWindow) {
-          resolve(iframe.contentWindow);
-        } else {
-          // We don't know of a case when this would happen, but better to fail
-          // fast if it does.
-          reject(
-            new Error(
-              `iframe.contentWindow not present on load for job "${jobId}".`,
-            ),
-          );
-        }
-      });
-
-      // We need to set the sandbox attribute after appending the iframe to the
-      // DOM, otherwise errors in the iframe will not be propagated via `error`
-      // and `unhandledrejection` events, and we cannot catch and handle them.
-      // We wish we knew why this was the case.
-      //
-      // We set this property after adding the `load` listener because it
-      // appears to work dependably. ¯\_(ツ)_/¯
-      //
-      // We apply this property as a principle of least authority (POLA)
-      // measure.
-      // Ref: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe#attr-sandbox
-      iframe.setAttribute('sandbox', 'allow-scripts');
-    });
   }
 }
