+++ packages/snaps-utils/src/index.ts
@@ -7,6 +7,7 @@ export * from './eval';
 export * from './flatMap';
 export * from './fs';
 export * from './handlers';
+export * from './iframe';
 export * from './json-rpc';
 export * from './manifest';
 export * from './mock';
